jp's suggestions + hopefully more digestible for code review
