+++ frontend/isoko/src/components/Home/KeywordSearchBar.tsx
@@ -4,7 +4,6 @@ import Autocomplete from '@mui/material/Autocomplete';
 import TextField from '@mui/material/TextField';
 import keywordList from '../../constants/keywordList';
 
-
 const StyledSearchBar = styled(Autocomplete)`
    border-right: 2px solid lightgray;
    background: #ffffff;
